requirements.txt
azure-ai-formrecognizer==3.1.2
azure-common==1.1.27
azure-core==1.21.0
azure-cosmos==4.2.0
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.1
Pillow==8.4.0
portalocker==1.7.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
